ML_project 1_Churn/requirements.txt
pandas
numpy
imblearn
scikit-learn
xgboost
plotly.express
seaborn
mlflow
dagshub
dvc
dvc-s3
python-dotenv